render as csv if result sets are too large to render as org-table
When iterating on a query, you may gather very large (i.e. thousands)
of results during predicate debugging.

You can specify what the largest amount of results you want to render
as an org-table is, anything above that will be returned as a raw
CSV file, but containing org links, such that code visiting still
works on the results.