prep for remote context use of both the lsp and query server
We need the args for LSP server invocation to be runtime configurable
so use a function to return the eglot params instead.

This is in prep for returning an alternate codeql-cli wrapper that
e.g. spawns on a remote host, running jsonrpc over stdio on top of
the remote connection.