Remove unneeded autoloads to prevent ql-tree-sitter-mode overload
ql-tree-sitter-mode does not need to be autoloadable as it is
always loaded from emacs-codeql, which will pick either builtin
or external treesitter support pending which emacs version is
in use.

This prevents overloading the autoload for ql-tree-sitter-mode
for folks that prefer to install emacs-codeql via a package.el
based interface.

Closes #7